+++ src/builtin.h
@@ -1145,6 +1145,23 @@ YDSH_METHOD module_dir(RuntimeContext &ctx) {
   RET(obj.getCode().getConstPool()[CVAR_OFFSET_SCRIPT_DIR]);
 }
 
+//!bind: function load($this : Module, $expr : String) : Func<Option<Any>>
+YDSH_METHOD module_load(RuntimeContext &ctx) {
+  SUPPRESS_WARNING(module_load);
+
+  auto &type = ctx.typePool.get(LOCAL(0).getTypeID());
+  auto ref = LOCAL(1).asStrRef();
+  assert(type.isModType());
+  auto &modType = cast<ModType>(type);
+  auto ret = loadExprAsFunc(ctx, ref, modType);
+  if (ret) {
+    RET(DSValue(ret.asOk()));
+  } else {
+    ctx.throwObject(DSValue(ret.asErr()), 1);
+    RET_ERROR;
+  }
+}
+
 // ###################
 // ##     Array     ##
 // ###################